# docs/swagger.yaml
basePath: /api
definitions:
  controller.StorageConfig:
    properties:
      s3_access_key:
        type: string
      s3_address:
        type: string
      s3_bucket:
        type: string
      s3_secret_key:
        type: string
      s3_token:
        type: string
      storage_base_path:
        type: string
      storage_type:
        type: string
    type: object
...
  domain.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Package Management System API for managing projects, packages, and
    users with RBAC.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PKMS API
  version: "1.0"
paths:
...
  /settings/storage:
    get:
      consumes:
      - application/json
      description: Get current storage configuration settings
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved storage configuration
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Get storage configuration
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Update storage configuration settings
      parameters:
      - description: Storage configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/controller.StorageConfig'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated storage configuration
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad request - invalid parameters
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Update storage configuration
      tags:
      - Settings
  /settings/storage/test:
    post:
      consumes:
      - application/json
      description: Test the current storage configuration connectivity
      produces:
      - application/json
      responses:
        "200":
          description: Storage connection test successful
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Unknown storage type or test failed
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Test storage configuration
      tags:
      - Settings
  /share/{code}/download:
    get:
      consumes:
      - application/json
      description: Download a release file using a share code without authentication
      parameters:
      - description: Share code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Successfully downloaded shared release file
          schema:
            type: file
        "404":
          description: Share not found or expired
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Download failed
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Download shared release
      tags:
      - Sharing
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"